fix: issue with embed
